refactor: rate shared
