+++ src/console/src/store/heap.rs
@@ -1,13 +1,15 @@
 use crate::memory::STATE;
 use crate::types::state::{
     Fee, Fees, HeapState, InvitationCode, InvitationCodeRedeem, InvitationCodes, MissionControls,
-    Payments, Rate, RateConfig, ReleaseVersion, ReleasesMetadata,
+    Payments, ReleaseVersion, ReleasesMetadata,
 };
 use ic_cdk::api::time;
 use ic_ledger_types::Tokens;
 use junobuild_shared::controllers::{
     delete_controllers as delete_controllers_impl, set_controllers as set_controllers_impl,
 };
+use junobuild_shared::rate::rate::increment_and_assert_rate;
+use junobuild_shared::rate::types::{Rate, RateConfig};
 use junobuild_shared::types::interface::SetController;
 use junobuild_shared::types::state::UserId;
 use junobuild_shared::types::state::{ControllerId, Controllers};
@@ -118,32 +120,17 @@ pub fn delete_controllers(remove_controllers: &[ControllerId]) {
 /// Rates
 
 pub fn increment_satellites_rate() -> Result<(), String> {
-    STATE.with(|state| increment_rate_impl(&mut state.borrow_mut().heap.rates.satellites))
+    STATE.with(|state| increment_and_assert_rate(&mut state.borrow_mut().heap.rates.satellites))
 }
 
 pub fn increment_mission_controls_rate() -> Result<(), String> {
-    STATE.with(|state| increment_rate_impl(&mut state.borrow_mut().heap.rates.mission_controls))
+    STATE.with(|state| {
+        increment_and_assert_rate(&mut state.borrow_mut().heap.rates.mission_controls)
+    })
 }
 
 pub fn increment_orbiters_rate() -> Result<(), String> {
-    STATE.with(|state| increment_rate_impl(&mut state.borrow_mut().heap.rates.orbiters))
-}
-
-fn increment_rate_impl(rate: &mut Rate) -> Result<(), String> {
-    let new_tokens = (time() - rate.tokens.updated_at) / rate.config.time_per_token_ns;
-    if new_tokens > 0 {
-        // The number of tokens is capped otherwise tokens might accumulate
-        rate.tokens.tokens = min(rate.config.max_tokens, rate.tokens.tokens + new_tokens);
-        rate.tokens.updated_at += rate.config.time_per_token_ns * new_tokens;
-    }
-
-    // deduct a token for the current call
-    if rate.tokens.tokens > 0 {
-        rate.tokens.tokens -= 1;
-        Ok(())
-    } else {
-        Err("Rate limit reached, try again later.".to_string())
-    }
+    STATE.with(|state| increment_and_assert_rate(&mut state.borrow_mut().heap.rates.orbiters))
 }
 
 pub fn update_satellites_rate_config(config: &RateConfig) {
