+++ src/libs/shared/src/rate/mod.rs
@@ -1,0 +1,3 @@
+pub mod constants;
+pub mod rate;
+pub mod types;
